% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{anomSP}
\alias{anomSP}
\title{Health anomaly between individual health and population median health not NA after deathtime}
\format{A matrix of dimensions 696 by 564; each row contains the estimates
of health anomaly for an individual. Each column is a month.}
\usage{
anomSP
}
\description{
This is a matrix of differences between individual health 
and population level health. I calculate this anomaly by first 
calculating the median health of the target groups (Adult males,
and young and old juveniles), and then subtracting that value
from the median health of each individual. Difference from \code{anom}
is that here I do \emph{not} set the health to NA after the month
in which the animal is estimated to have died
}
\keyword{datasets}

