% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAnom.R
\name{calcAnom}
\alias{calcAnom}
\title{Return the differences in health between individuals and pop average}
\usage{
calcAnom(ind, pop)
}
\arguments{
\item{ind}{- a 2 by t data frame of time-stamped individual health}

\item{pop}{- a 3 by T data frame of time-stamped population health with SD}
}
\value{
hdiff - the health anomaly
}
\description{
Goal of this is to take in the individual health data, along with the
population health data, and calculate the anomaly between the two
}

