% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKMCurvesSevGen.R
\name{calcKMCurvesSevGen}
\alias{calcKMCurvesSevGen}
\title{Return lists to be used in plotting Kaplan-Meier survival curves by gender and entanglement severity}
\usage{
calcKMCurvesSevGen(survdf, kdpasurvldf, nboot = 1, dcut, increment = 12,
  medProb = TRUE)
}
\arguments{
\item{\code{survdf}}{is the data frame containing survival data for all
animals in the three possible categories.}

\item{\code{kdpasurvldf}}{is the data frame containing survival data for
just the known dead and presumed alive animals}

\item{\code{nboot}}{an integer indicating the number of times we want to
run a bootstrap to collect uncertainty around the survivorship. Default
is to not run a boot strap}

\item{\code{dcut}}{integer indicating the month-year combo when the sightings 
data are considered complete through}

\item{\code{increment}}{the unit of time overwhich we calculate survivorship
Default value is to calculate it over a yearly interval}

\item{\code{medProb}}{A logical indicating whether or not the median estimate
of death is returned from the \code{getDeaths()} function. If TRUE, yes. 
If FALSE, then a value sampled from the posterior with normalised
probability for all candidate death months}
}
\value{
A list with two elements: 1) \code{kmlines} and 2) \code{censTicks}.
    The first element is a data frame of the survivorship for the different
    time frames. The second element is a data frame of censor times - simply
    used to make tick marks of when the animal was removed from the study
    via censoring.
}
\description{
This function is very similar to \code{calcKMCurves()} in that it takes the 
output from the functions used to make the survival data for the known 
dead, presumed alive, and presumed dead animals, and prepare the survivorship
information needed to plot the KM curves. Like that function, we can vary 
the interval over which the survivorship is calculated. And we can store 
uncertainty around the death times through the use of a boot strap. Where
it differs is that it is designed to examine survivorship split out
by gender and by entanglement severity.
}
\examples{
\dontrun{
calcKMCurvesSevGen(survdf, kdpasurvldf, nboot = 1, dcut, 
    increment = 12, medProb = TRUE)
}
}

