% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcpresdSurvdat.R
\name{calcpresdSurvdat}
\alias{calcpresdSurvdat}
\title{Prepare the data needed to calculate survivorship for presumed dead animals}
\usage{
calcpresdSurvdat(events, kdpasurvldf)
}
\arguments{
\item{\code{events}}{a data frame of the \emph{last} entanglement event
for individual whales.}

\item{\code{kdpasurvldf}}{a data frame output from the 
\code{calckdpaSurvdat()} function. These are the preparatory data for
calculating survivorship}
}
\value{
\code{survdf} a many by 10 stacked data frame containing the 
    information needed to calculate survivorship (Done in another function).
    \describe{
       \item{EGNo}{Numerical identifier of the individual whale}
       \item{deathMonth}{Integer noting the death month of the animal,
           i.e. the median month it was estimated to die in}
       \item{censored}{Logical indicating whether or not the animal was 
           censored at \code{dcut}}
       \item{censMonth}{Integer indicating the integer time for \code{dcut}}
       \item{survTime0}{Integer noting the length of time between the end 
           of the entanglement window and the \code{deathMonth}}
       \item{deathMonth0}{Integer noting the maximum of \code{survTime0}}
       \item{censMonth0}{Integer indicating number of months between
           the end of the entanglement and the censoring month 
           (\code{censMonth})}
       \item{Severity}{Character describing the severity of the entanglement
           injury. Ordinal values include: minor, moderate, and severe}
       \item{gearInj}{Integer denoting our numerical representation of the
           entanglement injury and whether or not the injury includes gear}
       \item{knownDeath}{Logical indicating whether or not the animal is
           known to have died}
    }
}
\description{
Goal of this is to ingest the \code{events} data frame and the output
from \code{calckdpaSurvdat()} to make one list that can be used
to calculate survivorship. The first element of this list will be
the presumed Alive and known Dead animals. All other elements will contain
the presumed dead animals. The goal of this structure was to be able
to
}
\examples{
\dontrun{
calcpresdSurvdat(events, kdpasurvldf)
}
}
