% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calfTable}
\alias{calfTable}
\title{Table describing all of the calving events in the right whale population}
\format{An object of class \code{data.frame} with 518 rows and 4 columns.}
\usage{
calfTable
}
\description{
This table is from Philip Hamilton (New England Aquarium) and contains
all the known information regarding calving. Each row contains calving 
information from the perspective of the mom. There are 4 columns in the data
frame as follows:
\describe{
   \item{EGNo}{4 number code that identifies individual right whale mothers}
   \item{CalvingYear}{Integer noting the year that the calf was born}
   \item{CalfNo}{The 4 number code identifying the ID of the calf. It's not
       known in all cases, e.g. '1973CalfOf1002'.}
   \item{CalvingInterval}{Integer describing the time since the mom last had a 
       calf}
}
}
\keyword{datasets}

