% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMerge.R
\name{cleanMerge}
\alias{cleanMerge}
\title{Clean up the entanglement time data for geary-carrying animals and merge it to \code{tangle}}
\usage{
cleanMerge(etime, tangle)
}
\arguments{
\item{etime}{- the timing of entanglement information for geary carrying 
animals.}

\item{tangle}{- the entanglement event data}
}
\description{
Goal of this is to pull in the entanglement time for gear carrying animals, 
and merge it with the entanglement event data for animals that are non-gear
carrying. The data come from Amy Knowlton, New England Aquarium.
}
\examples{
\dontrun{
cleanMerge(etime, tangle)
}
}

