% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{dcut}
\alias{dcut}
\title{Scalar noting when the sightings data are complete through}
\format{
A vector of length 1
}
\usage{
dcut
}
\description{
Right now this is set based on advice from Philip Hamilton (NEAq)
and sets the cutoff date for when the modelling 'ends.' The modelling
doesn't really end here, but this is the threshold we use for censoring
}
\keyword{datasets}
