% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{deadTable}
\alias{deadTable}
\title{Data frame of known deaths of individual right whales}
\format{A 42 by 8 data frame containing this columns:
\describe{
   \item{SightingEGNo}{Integer of the unique identifier of the individual whale}
   \item{SightingYear}{integer Year of death}
   \item{SightingMonth}{Integer Month of death}
   \item{SightingDay}{Integer Day of death}
   \item{Latitude}{Double Latitude position of observed death}
   \item{Longitude}{Double Longitude position of observed death}
   \item{AreaCode}{Character noting the macro-scale geographic location of the death}
   \item{RegionCode}{Character noting the micro-scale geographic location of the death}
}}
\usage{
deadTable
}
\description{
This comes from the New England Aquarium and is used to set the death time
and places for whales that are known to have died.
}
\keyword{datasets}

