% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{entvec}
\alias{entvec}
\title{Entanglement class definitions}
\format{A named vector, where the names are the character descriptions
and the vector values are integers}
\usage{
entvec
}
\description{
This is a length 6 named vector that describes the ordinal classes we
use for the entanglement injury types. There are 6 because we have 
three injury categories: minor, moderate, and severe along with whether
or not the animal is carrying gear.
}
\keyword{datasets}

