% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estStartDates}
\alias{estStartDates}
\title{Information about estimated start data of entanglement events}
\format{a data frame containing 130 rows and 6 variables:
\describe{
 \item{EntanglementId}{A unique identifier for the entanglement event in the NEAq DIGITS database}
 \item{EGNo}{4 number code that identifies individual right whales}
 \item{EventNo}{Integer labelling the consecutive number of each entanglement event}
 \item{StartDate}{A character variable describing the estimated first date of the 
 entanglement event}
 \item{EndDate}{A character variable describing the observed last date of the 
 entanglement event}
 \item{EntanglementComment}{Comments from Amy Knowlton about the entanglement events}
}}
\usage{
estStartDates
}
\description{
A dataset generated by Amy Knowlton of the New England Aquarium
as part of the PCAD/PCOD project looking at health and the impacts
of disturbance on vital rates in North Atlantic right whales. These
data are used to establish the start of "windows" that denote when animals 
are entangled and in turn to intersect those windows with estimates of health.
}
\keyword{datasets}

