% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etime}
\alias{etime}
\title{Information about gear carrying time on entangled right whales}
\format{
a data frame containing 102 rows and 9 variables: \describe{
  \item{EGNo}{4 number code that identifies individual right whales}
  \item{EventNo}{Integer labelling the consecutive number of each
  entanglement event} \item{Date prior w/no gear}{A character variable
  describing the first date the animal was seen without gear} \item{First
  date w/gear}{A character variable describing the first date the animal was
  seen with gear} \item{Last date w/gear}{A character variable describing the
  last date the animal was seen with gear} \item{line gone}{A character
  variable describing the first date the animal was seen that the gear is
  verified to be gone from the animal} \item{min time carrying gear}{Integer
  denoting the minimum number of days the animal was carrying the gear}
  \item{max dur}{Integer denoting the maximum number of days the animal was
  carrying the gear} \item{AmyEdits}{Comments from Amy Knowlton about
  individual entries she updated and/or inserted} }
}
\usage{
etime
}
\description{
A dataset generated by Amy Knowlton of the New England Aquarium as part of
the PCAD/PCOD project looking at health and the impacts of disturbance on
vital rates in North Atlantic right whales. These data are used to establish
the "windows" when animals are entangled and in turn to intersect those
windows with estimates of health.
}
\keyword{datasets}
