% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{firstSight}
\alias{firstSight}
\title{Vector denoting the first month (in integer form) of animal sighting}
\format{A vector of length 696}
\usage{
firstSight
}
\description{
Unlike \code{myName} this takes an integer form, e.g. 110, where 
\code{myName[firstSight[6]] == "2-1979"}. This is used throughout
the code to help extract values from the health and other matrices
}
\keyword{datasets}

