% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{gender}
\alias{gender}
\title{Gender of each individual right whale}
\format{A vector of length 696 comprised of 'M' for males and 'F' for females}
\usage{
gender
}
\description{
Gender of each individual right whale
}
\keyword{datasets}

