% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_subsequences}
\alias{get_subsequences}
\title{Title}
\usage{
get_subsequences(seq, ends = TRUE)
}
\arguments{
\item{seq}{amino acid sequence to get subsequences from}

\item{ends}{logical, if TRUE, adds the first and last amino acid with a "X" to the subsequences}
}
\description{
Title
}
