% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeaths.R
\name{getDeaths}
\alias{getDeaths}
\title{Matrix of imputed death times}
\usage{
getDeaths(deathyr, medProb = TRUE)
}
\arguments{
\item{deathyr}{A matrix where each row contains the estimated death times
for each animal. Times are in the columns from 1:bnt, which allows for
the animal to be alive at the time modelling end.}

\item{\code{medProb}}{logical indicating whether you want to sample
from the deaths using a multinomial, or just extract the median
death month (default)}
}
\value{
The output will be a vector of sampled death times. Each column
  will represent an individual animal's sampled times.
}
\description{
\code{getDeaths} returns a matrix of all of the estimated death times for each animal in the model
}
\details{
This is a function that will build an object to be used in the construction
of the Kaplan-Meier survival curves for right whales. The idea behind this
is to loop over the \code{deathyr} matrix and sample a death time for each
animal. The sampling is done with a call to \code{rmultinom} based on the 
estimates of death. These estimates are normalised, and then we choose a
candidate month.
}
\examples{
getDeaths(deathyr, medProb = TRUE)
}

