% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{healthmean}
\alias{healthmean}
\title{Estimated health of individual animals}
\format{A matrix of dimensions 696 by 564; each row contains the estimates
of health for an individual. Each column is a month.}
\usage{
healthmean
}
\description{
This is a matrix that of individual health estimated by the model
in Schick et al. 2013 PLoS-ONE.
}
\keyword{datasets}

