% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{healthmeanSP}
\alias{healthmeanSP}
\title{Estimated health of individual animals for slope plot}
\format{
A matrix of dimensions 696 by 564; each row contains the estimates
of health for an individual. Each column is a month.
}
\usage{
healthmeanSP
}
\description{
This is a matrix that of individual health estimated by the model
in Schick et al. 2013 PLoS-ONE. Difference from \code{healthmean}
is that here I don't set to NA the health for times after the estimated
death.
}
\keyword{datasets}
