% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{healthsd}
\alias{healthsd}
\title{Estimated standard deviation health of individual animals}
\format{A matrix of dimensions 696 by 564; each row contains the estimates
of the standard deviation around health for an individual. Each column is a 
month.}
\usage{
healthsd
}
\description{
This is a matrix of the standard deviation of individual health estimates 
returned by the model in Schick et al. 2013 PLoS-ONE.
}
\keyword{datasets}

