% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{ID}
\alias{ID}
\title{ID vector of EGNo's for all individual right whales in the data base}
\format{
A vector of length 696
}
\usage{
ID
}
\description{
This vector is generated and curated by the New England Aquarium. 
Each value in this vector is a unique and unchanging identifier 
of individual right whales
}
\keyword{datasets}
