% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{lastSight}
\alias{lastSight}
\title{Vector denoting the last month (in integer form) of animal sighting}
\format{A vector of length 696}
\usage{
lastSight
}
\description{
Unlike \code{myName} this takes an integer form, e.g. 110, where 
\code{myName[lastSight[6]] == "10-1986"}. This is used throughout
the code to help extract values from the health and other matrices
}
\keyword{datasets}

