% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_kmers}
\alias{make_kmers}
\title{Make all possible kmers of length n}
\usage{
make_kmers(n)
}
\arguments{
\item{n}{Length of kmers (3-6)}
}
\value{
A character vector of all possible kmers of length n
}
\description{
Make all possible kmers of length n
}
