% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_tangledDB}
\alias{make_tangledDB}
\title{Create the tangledDB database}
\usage{
make_tangledDB()
}
\description{
This function does not take any inputs. It creates the tangledDB database
that is also included as a dataset in the package.
}
