% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTangle.R
\name{makeTangle}
\alias{makeTangle}
\title{Create valid and usable variables describing the start/stop dates of entanglement windows}
\usage{
makeTangle()
}
\value{
\code{tangled} a data frame containing the cleaned entanglement data
}
\description{
The primary goal of this is to clean up the entanglement data a bit. We do this by
adding in information in the comments about each event, and coding that into a new
variable. We also extract the sex from the comments, and the age of the animal.
}

