% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeYearmon.R
\name{makeYearmon}
\alias{makeYearmon}
\title{to make a year format from monYr for the right whale plotting}
\usage{
makeYearmon(timedat)
}
\description{
Goal of this script is to make proper date-formatted year month
data from input time data
}
