% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{myName}
\alias{myName}
\title{Month-year vector used to mark time in the model}
\format{
A vector of length 696
}
\usage{
myName
}
\description{
This starts at January 1970, and goes through September, 2016,
or roughly the temporal extent of modelling. Values are like '1-1970'
}
\keyword{datasets}
