% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{ng}
\alias{ng}
\title{Number of Gibbs steps used in the input rda file}
\format{A vector of length 1}
\usage{
ng
}
\description{
This simply shows how long we ran the Gibbs model
for, and since we often store moments, this \code{ng}
object is useful to get values that make sense for 
subsequent analysis.
}
\keyword{datasets}

