% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{nt}
\alias{nt}
\title{Length of the modelling}
\format{A vector of length 1}
\usage{
nt
}
\description{
This is simply calculated with \code{length(myName)} and shows the 
temporal extent of the modelling.
}
\keyword{datasets}

