% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxplotHealth.R
\name{plotBoxplotHealth}
\alias{plotBoxplotHealth}
\title{Plot boxplots of health during entanglement and compare to unimpacted}
\usage{
plotBoxplotHealth(dfLong, bsize)
}
\arguments{
\item{\code{bsize}}{a scalar indicating how big to make the labels}

\item{\code{dfLong}}{a data frame contating health information
for both the reproductively active females and the 
\emph{non}-reproductively active females. These are the data for 
plotting}
}
\value{
A \code{ggplot2} object for plotting
}
\description{
Goal of this is to make the figure needed to depict median health 
during the entanglement windows as a function of reproductive status
and entanglement severity/class.
}
\examples{
\dontrun{
plotBoxplotHealth(dfLong, bsize)
}
}

