% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHealthThreshold.R
\name{plotHealthThreshold}
\alias{plotHealthThreshold}
\title{Make barplot of number of months below health threshold}
\usage{
plotHealthThreshold(dfb, bsize)
}
\value{
A \code{ggplot2} object for plotting
}
\description{
Goal is to simply make a bar plot of the number of months
animals are below certain health threshold values as a functino
of entanglement class and severity
}
\examples{
\dontrun{
plotHealthThreshold(dfb, bsize)
}
}

