% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperData.R
\docType{data}
\name{pophealth}
\alias{pophealth}
\title{Population level health of right whales}
\format{A data frame with 564 rows and 4 columns:
\describe{
\item{time}{Integer index of the month for the current row}
\item{popHealth}{Median population level health estimate}
\item{hsd}{Standard deviation of population level health estimate}
\item{date}{Date in \code{as.Date()} format of the month corresponding
    to the estimate.}
}}
\usage{
pophealth
}
\description{
This is a 4-column data frame containing information about the population
level health of the North Atlantic right whale population. It's calculated
using health information from Adult males, and both male and female 
juveniles.
}
\keyword{datasets}

