% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepBoxplotHealthData.R
\name{prepBoxplotHealthData}
\alias{prepBoxplotHealthData}
\title{Prepare data for making the boxplots of health during entanglement}
\usage{
prepBoxplotHealthData(tangRepro, tangNonRepro, anomFlag = FALSE, thold = 67)
}
\arguments{
\item{\code{tangRepro}}{data frame with information on the temporal
extend of windows for entnanglement events of 
reproductively active animals}

\item{\code{tangNonRepro}}{data frame with information on the temporal
extend of windows for entnanglement events of 
\emph{non}-reproductively active animals}
}
\value{
\code{dfLongRepro} a melted data frame of the health of 
    reproductively active animals during entanglement windows

\code{dfLongNonRepro} a melted data frame of the health of 
    \emph{non}-reproductively active animals during entanglement windows
}
\description{
The goal of this is to make the box plots of the health during the entanglement window.
This uses two data frames: (tangRepro and tangNonRepro) which are stored 
in the package structure. The function also used the unimpacted reference 
cases that are returned from the `returnUnimpactedHealth.R` helper
function. This function returns two data frames: \code{dfLongRepro}, and
\code{dfLongNonRepro}
}
\examples{
\dontrun{
prepBoxplotHealthData(tangRepro, tangNonRepro, anomFlag = TRUE, thold = 67)
}
}
