% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepHealthThresholdPlotData.R
\name{prepHealthThresholdPlotData}
\alias{prepHealthThresholdPlotData}
\title{Assemble the data for plotting the barplot of \% below x-health}
\usage{
prepHealthThresholdPlotData(tmp, rfmthold)
}
\arguments{
\item{\code{tmp}}{A list containing the health information from
the unimpacted females. This includes: \code{healthnew}
\code{nmon}, \code{nmonThold}, and \code{pThold}}
}
\value{
\code{rfmthold} A 6 by 7 data frame summarising the 
    information about the number of months entangled
    whales are spending below certain health thresholds
}
\description{
Goal of this is to take outputs from two different functions: 
1) \code{prephThresDataUnImp.R} and 2) \code{prephThreshDataRepro.R}
and assemble a data frame to be used for plotting with the
\code{plotHealthThreshold.R} function
}
\examples{
\dontrun{
prepHealthThresholdPlotData(tmp, rfmthold)     
}
}

