% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prephThreshDataRepro.R
\name{prephThreshDataRepro}
\alias{prephThreshDataRepro}
\title{Assemble health of impacted reproductively active females}
\usage{
prephThreshDataRepro(healthmean, thold)
}
\arguments{
\item{\code{healthmean}}{Estimates of health from the model}

\item{\code{thold}}{A scalar value of the health threshold to be
examined.}
}
\value{
A list containing: 1) \code{rfmthold} A 6 by 6 data frame 
    summarising the information about the number of months entangled
    whales are spending below certain health thresholds, and 2)
    \code{dfout} a 113 x 5 data frame containing the threshold
    data at the level of the infidividual, i.e. not summarised like
    in \code{rfmthold}
\describe{
  \item{gearInj}{The entanglement injury class of the animal. 
      1 = Severe w/Gear; 2 = Moderate w/Gear; 3 = Severe no gear;
      4 = Minor w/Gear; 5 = Moderate no gear; 6 = Minor no gear}
  \item{n}{The number of cases in each of these categories}
  \item{totNumMonths}{The total number of months tallied up
      from each individual whale, i.e. total number of
      months all whales were entangled}
  \item{maxDurMonths}{Maximum length of any entanglement}
  \item{sumBeltholdMonths}{Sum of the number of months
      entangled whales were below the threshold}
  \item{meanpthold}{The mean (by entanglement) percentage 
      of months the animal was below a threshold value
      during the event}     
}
}
\description{
This function takes heatlh estimates and intersects them
with the entanglement information to find out the number
of months each animal spent below a certain health threshold
}
\examples{
\dontrun{
prephThreshDataRepro(healthmean, thold)     
}
}

