% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prephThreshDataUnImp.R
\name{prephThreshDataUnImp}
\alias{prephThreshDataUnImp}
\title{Prepare data frame of health in unimpacted females for \% below barplot}
\usage{
prephThreshDataUnImp(healthmean, firstSight, lastSight, thold)
}
\arguments{
\item{\code{healthmean}}{A 696 by 564 matrix of estimated health
for all animals in the catalog dataset}

\item{\code{firstSight}}{A length 696 vector of the first month
each animal was seen}

\item{\code{lastSight}}{A length 696 vector of the last month
each animal was seen}

\item{\code{thold}}{A scalar value of the health threshold to be
examined.}
}
\value{
\code{healthnew} a 23 by 564 matrix of health estimates for 
    unimpacted females calving females before they've had their first
    calf, that is we don't want to examine health after the animals
    have had a calf and are subject to natural fluctuations related
    to the calving and nursing cycle.

\code{nAnimal} the number of unimpacted females
}
\description{
Goal is to look at the amount of time (in months) the whales are spending
below certain health thresholds. The idea is to see if female whales
are spending times in periods of low health and as a result have 
delayed or impaired fecundity. This function returns the health of 
these animals
}
\examples{
\dontrun{
prephThreshDataUnImp(healthmean, firstSight, lastSight, thold)
}
}

