% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSlopeHealthData.R
\name{prepSlopeHealthData}
\alias{prepSlopeHealthData}
\title{Prepare the individual line data for slope health plot}
\usage{
prepSlopeHealthData()
}
\value{
\code{dfout} A data frame containing the health and health anomaly 
for each entanglement event at each of the three locations
}
\description{
This function starts with the health data generated by the model
takes the entanglement event window, and using these to create
a data object summarising the change from start to finish of
an entanglement window. The input values of health are: 
\code{healthmeanSP} and \code{anomSP}, which are the estimated
health of each animal, and the anomaly of each animal's health
compared to the population median
}
\examples{
\dontrun{
prepSlopeHealthData()
}
}
