% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSlopeHealthDataAnnotation.R
\name{prepSlopeHealthDataAnnotation}
\alias{prepSlopeHealthDataAnnotation}
\title{Prepare the data for annotating the slope health plot}
\usage{
prepSlopeHealthDataAnnotation(dfout)
}
\value{
\code{dfn} A data frame containing the tally of the number of
events in each entanglement type.
}
\description{
This function starts with the \code{dfout} data frame, and puts that
through a \code{dplyr} wringer to calculate the number of cases
in each of the 6 different entanglement types.
}
\examples{
\dontrun{
prepSlopeHealthDataAnnotation(dfout)
}
}

