% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSlopeHealthDataBarplot.R
\name{prepSlopeHealthDataBarplot}
\alias{prepSlopeHealthDataBarplot}
\title{Prepare the data for barplot portion of the slope health plot}
\usage{
prepSlopeHealthDataBarplot(dfout)
}
\value{
\code{dfAsuml} A data frame containing the difference health 
in health anomaly in the two windows for each of the 6 conditions, 
i.e. health during and after for each of 6 entanglement types
}
\description{
This function starts with the \code{dfout} data, which is prepared
in \code{prepSlopeHealthData.R} and calculates the change during 
two different periods: 1) the entanglement window; and 2) the
recovery window. The goal is to show the magnitude of change
that is depicted with each different line plot in the upper
part of the graph.
}
\examples{
\dontrun{
prepSlopeHealthDataBarplot(dfout)
}
}
