% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSlopeHealthDataMedian.R
\name{prepSlopeHealthDataMedian}
\alias{prepSlopeHealthDataMedian}
\title{Prepare the median line data for slope health plot}
\usage{
prepSlopeHealthDataMedian(dfout)
}
\value{
\code{dfoutMed} A data frame containing the median health and 
health anomaly for each entanglement event at each of the three 
locations
}
\description{
This function starts with \code{dfout} and calculates median summary
lines to be used in the plotting of the upper panel of the slope plot
}
\examples{
\dontrun{
prepSlopeHealthDataMedian(dfout)
}
}
