% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presDeadsurvdat.R
\name{presDeadsurvdat}
\alias{presDeadsurvdat}
\title{Data frame of survival information for presumed dead animals}
\usage{
presDeadsurvdat(events, dcut, newDeadtimes)
}
\arguments{
\item{events}{A data frame of the entanglement data: timing, severity, etc.}

\item{dcut}{An numerical index between 1 and \code{nt} that is used to determine 
when animals are censored.}

\item{newDeadtimes}{a named vector of estimated death times. This is used to update
the events data, so that new survival information can be calculated.}
}
\value{
The output will be a data frame of survival information based on the
imputed death times.
}
\description{
\code{presDeadsurvdat} returns a matrix of all of the estimated death times for each animal in the model
}
\details{
This is a function that will build an object to be used in the construction
of the Kaplan-Meier survival curves for right whales. The idea behind this
is to loop over each animal that is presumed dead and calculate the
necessary informaiton for the survival curve plotting.

It will take in a new vector of death times (\code{newDeadtimes}), and replace the median
death times in \code{events} with the newly sampled ones.
}
\examples{
\dontrun{
presDeadsurvdat(events, dcut, newDeadtimes)
}
}

