% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnHealth_by_calving_interval.R
\name{returnHealth_by_calving_interval}
\alias{returnHealth_by_calving_interval}
\title{Return the health data for calving females during intervals}
\usage{
returnHealth_by_calving_interval(anomFlag = FALSE)
}
\value{
XXX
}
\description{
Calculate health by interval...
}
\examples{
\dontrun{
returnHealth_by_calving_interval()
}
}
