% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnhealthAnom.R
\name{returnhealthAnom}
\alias{returnhealthAnom}
\title{Return the right whale population level health Anomaly as a data frame}
\usage{
returnhealthAnom(healthmean)
}
\arguments{
\item{ngg}{A scalar corresponding to the number of Gibbs steps}

\item{tIndex}{A vector containing values 1:nt}

\item{monYr}{An nt by 2 matrix containing months in the first column and years in the second}

\item{sumh}{A matrix containing the estimates of right whale health for all whales}

\item{newgibbsMales}{A matrix containing the estimates of right whale health for Adult Males}

\item{newgibbsJuvs1}{A matrix containing the estimates of right whale health for Young Juveniles}

\item{newgibbsJuvs3}{A matrix containing the estimates of right whale health for Old Juveniles}
}
\value{
A data frame of right whale population level health anomaly, e.g. observed health (matrix) - pop health (vector)
}
\description{
Return the right whale population level health Anomaly as a data frame
}
\examples{
\dontrun{
returnhealthAnom(healthmean)
}
}

