% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPophealth.R
\name{returnPophealth}
\alias{returnPophealth}
\title{Return the right whale population level health as a data frame}
\usage{
returnPophealth()
}
\arguments{
\item{ngg}{A scalar corresponding to the number of Gibbs steps}

\item{tIndex}{A vector containing values 1:nt}

\item{monYr}{An nt by 2 matrix containing months in the first column and years in the second}

\item{newgibbsMales}{A matrix containing the estimates of right whale health for Adult Males}

\item{newgibbsJuvs1}{A matrix containing the estimates of right whale health for Young Juveniles}

\item{newgibbsJuvs3}{A matrix containing the estimates of right whale health for Old Juveniles}

\item{newgibbsMales2}{A matrix containing the estimates of standard deviation of right whale health for Adult Males}

\item{newgibbsJuvs12}{A matrix containing the estimates of standard deviation of right whale health for Young Juveniles}

\item{newgibbsJuvs32}{A matrix containing the estimates of standard deviation of right whale health for Old Juveniles}
}
\value{
A data frame of right whale population level health, comprised of median health between Adult males, and young and old juveniles
}
\description{
Return the right whale population level health as a data frame
}
\examples{
\dontrun{
returnPophealth()
}

}

