% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnSurvdf.R
\name{returnSurvdf}
\alias{returnSurvdf}
\title{Return survival data frame for testing differences in the curves}
\usage{
returnSurvdf(survdf)
}
\description{
We want to statistically test the differences between survival
as a function of gender and entanglement severity. So this function
will return an object formatted for use in the \code{survival}
package.
}
\examples{
\dontrun{
returnSurvdf(survdf)
}
}

