% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnUnimpactedHealth.R
\name{returnUnimpactedHealth}
\alias{returnUnimpactedHealth}
\title{Return the health data for unimpacted whales (reproductively active and non-reproductively active)}
\usage{
returnUnimpactedHealth()
}
\value{
A list element containing 2 data frames for health of the 1) 
   unimpacted and reproductive active whales and 2) unimpacted and 
   non-reproductive active whales.
}
\description{
This function creates and returns two data frames that will be used
in the plotting of health during the entanglement windows. IN particular
this function makes data frames that will be used as the reference cases
of health for non-impacted whales, i.e. what does the health look like
for animals that are not entangled. If a whale does at some point become
entangled, then we will only include their health (anomaly) estimates
up to the point of the first entanglement. This needs to be called
in order to make the boxplots
}
\examples{
\dontrun{
returnUnimpactedHealth()
}
}

