% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{substitution_matrix}
\alias{substitution_matrix}
\title{Substitution Matrix}
\usage{
substitution_matrix(matrix = "BLOSUM62")
}
\arguments{
\item{matrix}{A character string specifying the substitution matrix to use.}
}
\value{
A tibble with the substitution matrix.
}
\description{
Substitution Matrix
}
