% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tangleAll}
\alias{tangleAll}
\title{Information about entanglement events - raw}
\format{
A data frame of 1,194 rows and 12 columns:
\describe{
 \item{EntanglementId}{A unique identifier for the entanglement event in the NEAq DIGITS database}
 \item{EGNo}{4 number code that identifies individual right whales}
 \item{EventNo}{Integer labelling the consecutive number of each entanglement event}
 \item{StartDate}{A time variable describing the estimated first date of the 
 entanglement event}
 \item{EndDate}{A time variable describing the observed last date of the 
 entanglement event}
 \item{EntanglementComment}{Comments from Amy Knowlton about the entanglement events}
\item{AgeSex}{Character vector comment describing the age and gender of the animal, e.g. 'Unknown age female', '2 year old male', etc.}
\item{TimeFrame}{Amy Knowlton's descriptive comments concerning the time frame and duration of the entanglement event}
\item{Severity}{One of three categories of worsening injury: minor, moderate, severe. These are standard definitions created by NEA}
\item{gear}{Integer variable describing whether or not the animal is carrying gear as part of the entanglement: 0 for no gear; 1 for gear}
\item{LiveStatus}{Character vector indicating whether the animal is 'ALIVE' or 'DEAD'}
\item{Gender}{Character descriptor of the animal's sex: "female," "male," or NA}
}
}
\usage{
tangleAll
}
\description{
Information about entanglement events - raw
}
\keyword{datasets}
