% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winDiagram.R
\name{winDiagram}
\alias{winDiagram}
\title{Create a diagram showing how we extract health anomaly information during entanglement}
\usage{
winDiagram(anom, tangleOut, myName, egno = 1167)
}
\value{
A ggplot2 object for plotting the extraction diagram
}
\description{
Goal of this function is to return a ggplot2 plot object to show the extraction 
mechanics of intersecting the entanglement windows along with the health
information - in this case the health anomaly
}
\examples{
\dontrun{
winDiagram(anom, tangleOut, myName, egno = 1167)
}
}

