% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_TuNs}
\alias{get_TuNs}
\title{Title}
\usage{
get_TuNs(AA, type = "df")
}
\arguments{
\item{AA}{di-amino acid to search for in tangledDB}

\item{type}{type of return value, either "df" for a data frame or "vec" for a vector of unique TuNs}
}
\description{
Title
}
